Train/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
dotmap=1.3.26
numpy=1.22
pytorch=1.10.0
scikit-learn=1.0.1
scipy=1.7.3
tqdm=4.62.3
yaml=0.2.5
zlib=1.2.11
eigency=1.77
Cython=0.2.15